Extract useDragToClose hook to eliminate duplicated drag-to-close logic
Both ToDoItemPreviewDrawer and BirthdayPreviewDrawer had identical
implementations of pointer-based drag-to-close behavior. Extracted into
a reusable useDragToClose hook that accepts onClose and an optional
threshold parameter.
